server/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
